+++ agent/src/main/java/com/cmclinnovations/agent/utils/StringResource.java
@@ -28,10 +28,9 @@ public class StringResource {
       + "PREFIX ontoservice: <https://www.theworldavatar.com/kg/ontoservice/>"
       + "PREFIX xsd:<http://www.w3.org/2001/XMLSchema#>";
 
-  public static final String NODE_GROUP_VAR = "nodegroup";
-  public static final String NESTED_CLASS_VAR = "nested_class";
   public static final String CLAZZ_VAR = "clazz";
   public static final String RDF_TYPE = "rdf:type";
+  public static final String REPLACEMENT_PLACEHOLDER = "[replace]";
 
   // Private constructor to prevent instantiation
   private StringResource() {
@@ -88,6 +87,15 @@ public class StringResource {
    */
   public static String genOptionalClause(String content) {
     return "OPTIONAL{" + content + "}";
+  }
+
+  /**
+   * Wraps the content into {}
+   * 
+   * @param content Target content
+   */
+  public static String genGroupGraphPattern(String content) {
+    return "{" + content + "}";
   }
 
   /**